mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.74.0",
    "openai>=1.50.0",
    "mcp>=1.21.2",
    "python-dotenv>=1.2.1",
    "beautifulsoup4>=4.13.3",
    "tiktoken>=0.9.0",
]
